twoliter: add encrypted-storage image feature
Introduce a new, experimental image feature to encrypt local storage.
This is only compatible with platforms that provide a TPM2 device, so
that the required keys can be protected.

To avoid trivial offline attacks, where encrypted storage could be
replaced by unencrypted data in the expected format, any policy for
encryption needs to be defined at the variant level and then enforced
unconditionally by the OS.

As with most other image features, enabling this flag for an existing
variant is backwards-incompatible, and migrations into or out from an
encrypted state are not supported.

Signed-off-by: Ben Cressey <bcressey@amazon.com>