twoliter: record in-place updates state in image-features.env
This allows programs to query whether in-place updates are enabled
without relying on other indicators that might change over time, such
as whether certain partitions are present.

In the specific case of encrypted storage, when in-place updates are
disabled, we can use more of the PCRs to seal encryption keys, since
the kernel and userspace should never change. Since this decision has
to be made before the API server's datastore can be created, we can't
rely on an API to discover the feature state.

Signed-off-by: Ben Cressey <bcressey@amazon.com>